refactor: Generalized the Hero component
