+++ src/Main/components/Common/Hero.jsx
@@ -1,7 +1,4 @@
 import React from 'react'
-import GOONJ from '../../assets/GOONJ.svg'
-import Cross from '../../assets/cross.svg'
-import GDSC from '../../assets/GDSC.svg'
 import circle from '../../assets/shapes/circle.svg'
 import circle2 from '../../assets/shapes/circle-2.svg'
 import ellipse from '../../assets/shapes/ellipse.svg'
@@ -9,7 +6,7 @@ import ellipse2 from '../../assets/shapes/ellipse-2.svg'
 import squarePlate from '../../assets/shapes/square-bg.svg'
 
 
-const Hero = () => {
+const Hero = ({ children, text }) => {
     return (
         <section className='mt-5 lg:mt-0'>
             <div className='flex items-start justify-between'>
@@ -17,11 +14,7 @@ const Hero = () => {
                     <img src={squarePlate} alt="plate" height={282} width={287} />
                     <img src={circle} alt="circle" width={60} />
                 </div>
-                <div className='flex items-center justify-center gap-3 lg:gap-28 mt-28'>
-                    <img src={GDSC} alt="GDSC" height={100} width={200} />
-                    <img src={Cross} alt="Cross" height={50} width={50} />
-                    <img src={GOONJ} alt="GOONJ" height={150} width={200} />
-                </div>
+                {children}
                 <div className='relative'>
                     <img src={squarePlate} alt="plate" width={282} height={282} />
                     <img src={ellipse2} alt="ellipse" width={60} className='absolute top-10 left-20' />
@@ -34,7 +27,7 @@ const Hero = () => {
                         <img src={squarePlate} alt="" height={400} width={400} />
                         <img src={ellipse} alt="ellipse" width={60} className='absolute top-0 left-80' />
                         <img src={circle2} alt="circle" width={60} className='absolute top-10 right-64' />
-                        <p className='absolute w-full text-5xl font-bold top-[calc(50%-30px)] left-[calc(50%-180px)] text-left'>13, 14 and 15 March</p>
+                        <p className='absolute w-full text-5xl font-bold top-[calc(50%-30px)] left-[calc(50%-180px)] text-left'>{text}</p>
                     </div>
                 </div>
             </div>
